requirements/base.txt
# basic requirements
django
djangorestframework
django-cors-headers
